feat: remove deprecation warning for nightly yazi users
Issue
=====

The nightly version of yazi has changed the way arguments are passed to
the plugin entry point. Previously they were a table that contained the
arguments, but they were moved inside {args = {...}}.

If the user is using a nightly version of yazi, they will see a
deprecation warning.

https://github.com/sxyazi/yazi/pull/1966

Solution
========

This change removes the warning and supports both versions.